// packages/typegpu/src/future/types.ts
export interface WgslResolvable {
  readonly label?: string | undefined;

  resolve(ctx: ResolutionCtx): string;
}
// ...
export interface WgslAllocatable<TData extends AnyWgslData = AnyWgslData> {
  /**
   * The data type this allocatable was constructed with.
   * It informs the size and format of data in both JS and
   * binary.
   */
  readonly dataType: TData;
  vertexLayout: Omit<GPUVertexBufferLayout, 'attributes'> | null;
  readonly initial?: Parsed<TData> | WgslPlum<Parsed<TData>> | undefined;
  readonly flags: GPUBufferUsageFlags;
  get label(): string | undefined;
}
// ...
export type BufferUsage = 'uniform' | 'readonly' | 'mutable' | 'vertex';
// ...
export interface WgslData<TInner> extends ISchema<TInner>, WgslResolvable {
  readonly byteAlignment: number;
  readonly size: number;
}

export type AnyWgslData = WgslData<unknown>;